requirements.txt
python-binance~= 1.0.16
requests~=2.27.1
config~=0.5.1
SQLAlchemy~= 2.0.0